LPS-180450 Use the pattern for DB Partition tools
