+++ modules/util/portal-tools-db-partition-virtual-instance-migration/src/main/java/com/liferay/portal/tools/db/partition/virtual/instance/migration/error/ErrorCodes.java
@@ -1,0 +1,38 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.tools.db.partition.virtual.instance.migration.error;
+
+/**
+ * @author Luis Ortiz
+ */
+public class ErrorCodes {
+
+	public static final int BAD_DESTINATION_PARAMETERS = 12;
+
+	public static final int BAD_INPUT_ARGUMENTS = 10;
+
+	public static final int BAD_SOURCE_PARAMETERS = 11;
+
+	public static final int DESTINATION_NOT_DEFAULT = 13;
+
+	public static final int SOURCE_MULTI_INSTANCES = 14;
+
+	public static final int SUCCESS = 0;
+
+	public static final int UNEXPECTED_ERROR = 255;
+
+	public static final int VALIDATION_ERROR = 20;
+
+}
