--- modules/util/portal-tools-db-virtual-instance-migration/bnd.bnd
@@ -1,4 +1,0 @@
-Bundle-Name: Liferay Portal Tools DB Virtual Instance Migration
-Bundle-SymbolicName: com.liferay.portal.tools.db.virtual.instance.migration
-Bundle-Version: 1.0.0
-Main-Class: com.liferay.portal.tools.db.virtual.instance.migration.VirtualInstanceMigration
