--- modules/util/portal-tools-db-virtual-instance-migration/scripts/db_virtual_instance_migration.bat
@@ -1,11 +1,0 @@
-@echo off
-
-pushd "%~dp0"
-
-path %PATH%;%JAVA_HOME%\bin
-
-java -jar com.liferay.portal.tools.db.virtual.instance.migration.jar %*
-
-popd
-
-@echo on
