--- modules/util/portal-tools-db-virtual-instance-migration/source-formatter-suppressions.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0"?>
-
-<suppressions>
-	<checkstyle>
-		<suppress checks="CompanyIterationCheck" files="src/main/java/com/liferay/portal/tools/db/virtual/instance/migration/internal/util/Database\.java" />
-	</checkstyle>
-	<source-check>
-		<suppress checks="StringMethodsCheck" files="src/main/java/com/liferay/portal/tools/db/virtual/instance/migration/internal/util/Database\.java" />
-	</source-check>
-</suppressions>
